input: synaptics: s3320: Clean up optimized-init insanity
Offloading init onto a worker is a good idea to prevent slowing down
overall kernel init with our probe, but the way it's being done is rather
insane (especially with the CPU selection logic).

Clean all of this up by just putting the i2c_add_driver() into a worker.
Additionally, increase the probe delay from 300 ms to 1000 ms since some
users report the touchscreen not working on boot (so this is just for good
measure), and remove the unnecessary module references while we're at it.

Signed-off-by: Sultanxda <sultanxda@gmail.com>